Cleanup and appease detekt
