+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/details/manage/DeleteConfirmationDialogViewModel.kt
@@ -1,0 +1,10 @@
+package org.noiseplanet.noisecapture.ui.features.details.manage
+
+import org.jetbrains.compose.resources.StringResource
+
+data class DeleteConfirmationDialogViewModel(
+    val title: StringResource,
+    val text: StringResource,
+    val onDismissRequest: () -> Unit,
+    val onConfirm: () -> Unit,
+)
