+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/details/manage/ManageMeasurementMenuItem.kt
@@ -1,0 +1,9 @@
+package org.noiseplanet.noisecapture.ui.features.details.manage
+
+import org.jetbrains.compose.resources.StringResource
+
+data class ManageMeasurementMenuItem(
+    val label: StringResource,
+    val supportingText: StringResource?,
+    val onClick: () -> Unit,
+)
